Introducing Application Telemetry Schema in OpenTelemetry - Vision and Roadmap (#243)
[Easy to read version with images:
[https://github.com/lquerel/oteps/blob/app-telemetry-schema-vision-roadmap/text/0243-app-telemetry-schema-vision-roadmap.md](https://github.com/lquerel/oteps/blob/app-telemetry-schema-vision-roadmap/text/0243-app-telemetry-schema-vision-roadmap.md)]

Unlike the traditional data ecosystem (OLTP and OLAP), the world of
telemetry generally does not rely on the concept of a schema.
Instrumentation is deeply embedded in the code of applications and
libraries, making it difficult to discover all the possible telemetry
signals an application can emit. This gap prevents or limits the
development of CI/CD tools for checking, reporting, documenting, and
generating artifacts from telemetry signals specific to an application.
This document presents a long-term vision aimed at enabling the
OpenTelemetry project to address this issue and extend its impact to a
broader ecosystem. It proposes extending the initiatives of Telemetry
Schema and Semantic Conventions to include concepts of Application
Telemetry Schema and Resolved Telemetry Schema. A series of OTEPs and
Tools will be proposed in this overarching document to detail each
aspect of this vision.

Similar (but proprietary) initiative from Facebook: [Positional Paper:
Schema-First Application
Telemetry](https://research.facebook.com/publications/positional-paper-schema-first-application-telemetry/)

**EDIT 1: The OTel Weaver project (a PoC implementation of this OTEP and
some of the others mentioned in the roadmap) is now available
[here](https://github.com/f5/otel-weaver).**

**EDIT 2: The Slack channel
[#otel-weaver](https://cloud-native.slack.com/archives/C0697EXNTL3) is
dedicated to this OTEP and the associated OTel Weaver project.**

---------

Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>
Co-authored-by: Reiley Yang <reyang@microsoft.com>
Co-authored-by: Tigran Najaryan <4194920+tigrannajaryan@users.noreply.github.com>