Remove implementation detail from Logs Bridge API
The bridge-api.md document specifies the Logs Bridge API, it does not
defined the OpenTelemetry defined implementations of this API. Remove
this detail from the API and leave it for the implementation
specification to define.

Importantly, the API is decoupled from implementation for a reason.
Third-party implementation may exists that OTel does not have domain
over. This document needs to written with that in mind.