Update status of stdout exporter for logs to stable
https://github.com/open-telemetry/opentelemetry-specification/pull/3741/files and 
https://github.com/open-telemetry/opentelemetry-specification/pull/3740/files
were added together, and it is not clear why one is experimental, and the other is stable.

This PR is marking the logging stdout exporter spec as stable. I couldn't find any reason why it was non-stable in the first place, so I assume it was just an oversight only. Happy to be corrected, if I missed something