Prohibit attribute value from evolving to contain complex types (#3858)
If we aren't going to accept complex attribute types (#2888) we should
explicitly rule them out of future designs. Doing so cements the idea
that attributes are "metadata" instead of "data", since if attributes
were data, we would not want to artificially limit their structure. Once
its clear that attributes are metadata and restricted to a limited set
of types, its easy to determine that use cases which require complex
types (like event payloads) should seek to put the data elsewhere (like
in a log record body).

While I was in favor of supporting complex attribute types (#2888) I
believe its more important that we commit one way or the other. The
uncertainty around the question of whether this type of evolution will
occur has muddied the waters of several related conversations.

There was consensus on codifying this in the 1/30/24 spec SIG meeting.
We should capitalize on this momentum and get this over the finish line.
Stalling out just to revisit this same debate in the future is a bad use
of time.