Clarify differences between prometheus protocols, including target info metric naming (#3871)
Fixes
https://github.com/open-telemetry/opentelemetry-specification/issues/3700

## Changes

The current specification mixes OpenMetrics and Prometheus in ways that
can be confusing. This change first explains the differences between
various Prometheus formats, and then consistently uses "Prometheus" to
refer to the general conversion rules. This is helpful for explaining
why the target (info) metric naming differs depending on the format
used.

Also, fix references to the "target" info-typed metric, which was
previously referred to as "target_info". "target_info" is only used when
info-typed metrics are not supported.

---------

Co-authored-by: Yuri Shkuro <yurishkuro@users.noreply.github.com>
Co-authored-by: Eero Tamminen <eero.t.tamminen@intel.com>
Co-authored-by: Robert Pająk <pellared@hotmail.com>
Co-authored-by: Anthony Mirabella <a9@aneurysm9.com>
Co-authored-by: Josh Suereth <joshuasuereth@google.com>