Introduce Entities Data Model, Part 1 (open-telemetry/oteps#256)
This is a proposal of a data model to represent entities. The purpose of
the data model is to have a common understanding of what an entity is,
what data needs to be recorded, transferred, stored and interpreted by
an entity observability system.

This data model sets the foundation for adding entities to
OpenTelemetry. The data model is largely borrowed from
[the initial
proposal](https://docs.google.com/document/d/1VUdBRInLEhO_0ABAoiLEssB1CQO_IcD5zDnaMEha42w/edit)
that was accepted for entities SIG formation.

This OTEP is step 1 in introducing the entities data model. Follow up
OTEPs will add further data model definitions, including the linking of
Resource information to entities.