Rename Experimental to Development according to OTEP 0232 (#4061)
Resolves
https://github.com/open-telemetry/opentelemetry-specification/issues/4051

[OTEP
0232](https://github.com/open-telemetry/oteps/blob/main/text/0232-maturity-of-otel.md)
brought new stability definitions.

The impact on this repository is that we need to rename "Experimental"
to "Development". All other existing statuses remain unchanged.

Note: OTEP 0232 also introduces other levels, such as Alpha, Beta, etc.
We will need to decide separately if we want to start using these levels
in this repository.