Stabilize trace context in non-OTLP formats and define casing (#3909)
Closes
https://github.com/open-telemetry/opentelemetry-specification/issues/3303

## Changes

From @tigrannajaryan in
https://github.com/open-telemetry/opentelemetry-specification/issues/3303#issuecomment-1967252364:

> I think the conclusion in
https://github.com/open-telemetry/opentelemetry-specification/issues/3518
was that we keep the names. It has been a while and there is no new
evidence that we need to do something else so I suggest that we submit a
PR that declares
https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/compatibility/logging_trace_context.md
"Stable" and closes this issue.

This marks the specification for using `trace_id`, `span_id`, and
`trace_flags` in non-OTLP logging formats stable.