Add ability for SpanProcessor to mutate spans on end (#4024)
Fixes #1089.

In addition to the comments on the issue, this was discussed in the spec
SIG Meeting on 2024/23/04:
* The filtering use-case explained in [this
comment](https://github.com/open-telemetry/opentelemetry-specification/issues/1089#issuecomment-2061289086)
should rather be solved by the upcoming samplerV2 instead of
`SpanProcessor`s due to better conceptual fit
* The buffering use-case also explained in [this
comment](https://github.com/open-telemetry/opentelemetry-specification/issues/1089#issuecomment-2061289086)
seems to be not relevant enough to influence the design decision
* Apparently there was a discussion around building the `SpanProcessor`s
in a chaining fashion during the initial SDK spec design and it was
actively decided against it. However, no one could recall the reason
why.