Allow cloning of ReadWriteLogRecord (#4090)
Follows
https://github.com/open-telemetry/opentelemetry-specification/pull/4062

## Why

1. There is a need to make a deep copy of a `ReadWriteRecord` in order
to implement an experimental isolating processor outside of the SDK (as
it is experimental, we prefer to land it to contrib repository first).
2. Allow fine-grained control for the users to make a deep copy only
when necessary. This allows users to have more control on the processing
e.g. pass a clone to asynchronous processor to avoid race conditions or
make performance improvements.