Add support for filtering attribute keys for streams via an exclude list (#4188)
## Changes

This change adds details around supporting the creation of a stream with
a set of excluded attribute keys. This functionality is supported in the
[Go
implementation](https://github.com/open-telemetry/opentelemetry-go/blob/main/sdk/metric/instrument.go#L146)
via the use of filters. The proposal in this PR is to address an issue
in the configuration working group:
https://github.com/open-telemetry/opentelemetry-configuration/issues/112

Please provide a brief description of the changes here.

For non-trivial changes, follow the [change proposal
process](https://github.com/open-telemetry/opentelemetry-specification/blob/main/CONTRIBUTING.md#proposing-a-change).

* [x] Links to the prototypes (when adding or changing features) 
* [x]
[`CHANGELOG.md`](https://github.com/open-telemetry/opentelemetry-specification/blob/main/CHANGELOG.md)
file updated for non-trivial changes
* [x]
[`spec-compliance-matrix.md`](https://github.com/open-telemetry/opentelemetry-specification/blob/main/spec-compliance-matrix.md)
updated if necessary

---------

Signed-off-by: Alex Boten <223565+codeboten@users.noreply.github.com>