Add requirement to allow extending Stable APIs
Resolves https://github.com/open-telemetry/opentelemetry-specification/issues/4257

This issue has been discussed in spec SIG meeting on 22 Oct 2024 and
decision was made that we want this to be a requirement for language
implementations.

This is a new requirement for implementations, which we believe
becomes more and more important now that we have Stable signals that
we would like to continue evolving.