Update explicit context injection example for Log Bridges (#4406)
Update the supplementary guidelines to refer to an example which is
actually implemented in OTel Go Contrib.

The "initial" example requires changing the "bridged" logging library.
Therefore, it is not an user-friendly bridge.

The actual implementation is inspired by @yurishkuro comment from the
doc:
https://docs.google.com/document/d/15vR7D1x2tKd7u3zaTF0yH1WaHkUr2T4hhr7OyiZgmBg/edit?disco=AAAAVOJM1u8.
However, we do not even require/define a special Field type that
captures the context.

`otelzap` was released Aug 23, 2024 and we have not received a single
complain about the existing design/behavior. We use the same pattern in
`otellogr`.