Add subscript for issue participation (#4550)
## Changes

As part of https://github.com/open-telemetry/sig-end-user/issues/11,
this adds a subscript paragraph to all issue templates to help end-users
and other community members to interact with issues.

This change is accompanied by the changes in the maintainer guide
proposed in https://github.com/open-telemetry/community/pull/2778, and
the changes to the website in
https://github.com/open-telemetry/opentelemetry.io/pull/7088 giving
users more information about issue participation

I propose we trial this in the spec repo and see if it's useful for
end-users and for us.

---------

Co-authored-by: Trask Stalnaker <trask.stalnaker@gmail.com>