OTEP: Extending attributes to support complex values (#4485)
Related to #4468, #4414

This OTEP suggests a path forward on how to support complex attributes.
It proposes to:

- allow complex attributes on all signals by default on the API and SDK
level
- discourage them on metrics, resources, instrumentation scope, and
identifying attributes on entities. If complex attribute is provided,
it's probably by mistake, but we let backend decide.
- provide guidance in semconv to default to flat attributes whenever
possible and develop conventions with an assumption that complex
attributes are not indexed, not queryable and not efficient.
- documents how limits can apply to complex attributes (leaf nodes are
counted towards the attribute count limit and are truncated based on the
value limit) and suggests other ways to protect from users misusing
complex attributes.

Having experimental phase for complex attributes limits the impact of
this change and gives backends a graceful period to update.

---------

Co-authored-by: Trask Stalnaker <trask.stalnaker@gmail.com>
Co-authored-by: Robert Pająk <pellared@hotmail.com>
Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>
Co-authored-by: Adrian Garcia Badaracco <1755071+adriangb@users.noreply.github.com>
Co-authored-by: Alan West <3676547+alanwest@users.noreply.github.com>
Co-authored-by: jason plumb <75337021+breedx-splk@users.noreply.github.com>