Improve concurrency safety description of LogRecordProcessor.OnEmit (#4578)
Fixes
https://github.com/open-telemetry/opentelemetry-specification/issues/4141

## Why

I do not think it is possible to disallow modifications of `logRecord`
after `LogRecordProcessor.OnEmit`. It is a "requirement" that is left
for the user.

## What

Instead of it I think we should explain the problem and also describe
ways how race conditions can be mitigated/solved.

This is how we implemented it in OTel Go:
- docs:
https://github.com/open-telemetry/opentelemetry-go/blob/69e81088ad40f45a0764597326722dea8f3f00a8/sdk/log/processor.go#L31-L38
(docs)
- mitigating a possible race conditions in the batch processor:
https://github.com/open-telemetry/opentelemetry-go/blob/48fedfa58e35b5f3b08fe54e45cfad441dd82426/sdk/log/batch.go#L185-L187

PS: Sorry it took me so long to create this PR (almost a year 😅). It
actually took me more than an hour just to figure out how to propose
these changes in a hopefully good way. Thank you for your patience!

---------

Co-authored-by: Tyler Yahn <MrAlias@users.noreply.github.com>
Co-authored-by: Carlos Alberto Cortez <calberto.cortez@gmail.com>