[Configuration SDK] Rename Config operations to SDK operations (#4587)
The current term `Config operations` is very confusing. 
Especially that there is a `ConfigProvider` which is NOT returning
objects that have "Config operations".

It looks that the operations defined in "Config operations" could be
"static/global" e.g. (prototypes):
-
https://pkg.go.dev/go.opentelemetry.io/contrib/otelconf/v0.3.0#ParseYAML
- https://pkg.go.dev/go.opentelemetry.io/contrib/otelconf/v0.3.0#NewSDK

Note that even the section starts with

> SDK implementations of configuration MUST provide the following
operations.

I also think that [the creation of
`ConfigProvider`](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/configuration/sdk.md#configprovider)
should be documented as another SDK operation, but I would rather scope
it to a separate PR.

Co-authored-by: Carlos Alberto Cortez <calberto.cortez@gmail.com>