Update YAML spec compliance matrix for Ruby (#4647)
## Changes

This PR is a redo of #4628 to account for the changes made in #4631. 

To quote the description of #4628: 
> The Ruby Metrics and Logs implementations added a lot of features
since this matrix was last updated. However, we're still in
"Development" status, so I'm not sure if this is the appropriate time to
update the matrix. Please let me know if we should wait.

This also includes some changes to the traces and environment variables
sections. There are a few small differences between the boxes checked in
#4628 to reflect recent changes.

cc @open-telemetry/ruby-maintainers

---------

Co-authored-by: Carlos Alberto Cortez <calberto.cortez@gmail.com>