Changes of disabled config MUST be eventually visible (#4645)
Fixes
https://github.com/open-telemetry/opentelemetry-specification/issues/4643

## What

Clarify that changes of disabled config MUST be eventually visible.

## Why

Currently, the implementation that would not never synchronize
modification and access to `disabled` config are compliant with the
specification. This is not correct as then the user may never see the
changes when the configuration changes.

Per
https://github.com/open-telemetry/opentelemetry-specification/pull/4645#discussion_r2334091832

Languages remain free to document and implement their own
refresh/caching behaviors as long as they are eventually consistent.

---------

Co-authored-by: Liudmila Molkova <neskazu@gmail.com>
Co-authored-by: Carlos Alberto Cortez <calberto.cortez@gmail.com>