Move experimental sampling specification into OTEPs (#4673)
Goal is to improve our public documentation, which has two documents
with the same title today, the second one being correct.

Finishes the deprecation process started in #4168.
When #4166 added the new sampling specification, the old one was
renamed.
Now that the new specification is complete in the specification, the old
one is a distraction.

## Changes

This moves the old document into the `oteps/` repository where it will
remain a permanent record of this effort.