Clarify declarative config default requirements (#4780)
When the declarative config `create` spec was written, declarative
config was dependent on the spec as its source of truth for default
values for the various properties.

We now directly record expected [`defaultBehavior` and
`nullBehavior`](https://github.com/open-telemetry/opentelemetry-configuration/blob/main/CONTRIBUTING.md#json-schema-source-and-output)
semantics in `opentelemetry-configuration`.

Updating the spec to tie things back together. 

/cc @open-telemetry/configuration-approvers